Fix Linux support
The plugin didn't work on Linux (and probably not on Mac either) because Popen expects a list containing the command and its arguments, rather than everything as a string separated by spaces. I fixed this, which also removed the need for shlex. I've only tested on Linux so please test it on Windows as well.